+++ src-tauri/meter-data/Zone.json
@@ -1,0 +1,90 @@
+{
+  "30801": "Frost Haven",
+  "30802": "Foggy Ridge",
+  "30803": "Red Sand Desert",
+  "30804": "Scarred Crimson Land",
+  "30805": "Woeful Drylands",
+  "30806": "Aberrant Mist Forest",
+  "30807": "Shadowed Swamp of Silence",
+  "30835": "Temple of Reconstructed Memory",
+  "37001": "Sea of Indolence",
+  "37002": "Tranquil Karkosa",
+  "37003": "Alaric's Sanctuary",
+  "37011": "Demon Beast Canyon",
+  "37012": "Necromancer's Origin",
+  "37021": "Hall of the Twisted Warlord",
+  "37022": "Hildebrandt Palace",
+  "37031": "Road of Lament",
+  "37032": "Forge of Fallen Pride",
+  "37041": "Aira's Oculus",
+  "37042": "Oreha Preveza",
+  "37051": "Resurrected Demon Beast Lair",
+  "37061": "Garden of Crimson Delight",
+  "37071": "Corrupt Lord's Pandemonium",
+  "37072": "Corrupt Lord's Pandemonium",
+  "37081": "Midnight Circus",
+  "37091": "Phantom Astalgia",
+  "37092": "Phantom Astalgia",
+  "37093": "Phantom Astalgia",
+  "37094": "Phantom Astalgia",
+  "37101": "Eternal Cradle of Light",
+  "37102": "Eternal Cradle of Light",
+  "37111": "Trampled Garden",
+  "37112": "Trampled Garden",
+  "37121": "Dark Baratron",
+  "37122": "Dark Baratron",
+  "37123": "Dark Baratron",
+  "37124": "Dark Baratron",
+  "308010": "",
+  "308011": "",
+  "308012": "",
+  "308014": "",
+  "308015": "",
+  "308016": "",
+  "308017": "",
+  "308018": "",
+  "308019": "",
+  "308020": "",
+  "308021": "",
+  "308022": "",
+  "308023": "",
+  "308024": "",
+  "308025": "",
+  "308026": "",
+  "308027": "",
+  "308028": "",
+  "308029": "",
+  "308030": "",
+  "308037": "",
+  "308039": "",
+  "308040": "",
+  "308041": "",
+  "308042": "",
+  "308043": "",
+  "308044": "",
+  "308239": "",
+  "308339": "",
+  "308410": "",
+  "308411": "",
+  "308412": "",
+  "308414": "",
+  "308415": "",
+  "308416": "",
+  "308417": "",
+  "308418": "",
+  "308419": "",
+  "308420": "",
+  "308421": "",
+  "308422": "",
+  "308423": "",
+  "308424": "",
+  "308425": "",
+  "308426": "",
+  "308428": "",
+  "308429": "",
+  "308430": "",
+  "308437": "",
+  "309020": "",
+  "30865": "Neglected Ancient Ruin",
+  "30866": "Behemoth's Hideout"
+}
