+++ src-tauri/src/context.rs
@@ -1,0 +1,39 @@
+use std::path::PathBuf;
+
+use anyhow::{Result, ensure};
+use single_instance::SingleInstance;
+
+use crate::constants::*;
+
+#[derive(Debug, Clone)]
+pub struct AppContext {
+    pub app_path: PathBuf,
+    pub current_dir: PathBuf,
+    pub settings_path: PathBuf,
+    pub database_path: PathBuf,
+    pub local_player_path: PathBuf,
+    pub region_file_path: PathBuf,
+}
+
+impl AppContext {
+    pub fn new() -> Result<Self> {
+        let instance = SingleInstance::new("LOA_LOGS").unwrap();
+        ensure!(instance.is_single(), "Another instance of {} is already running!", "LOA Logs");
+
+        let app_path = std::env::current_exe().unwrap();
+        let current_dir = app_path.parent().unwrap().to_path_buf();
+        let settings_path = current_dir.join(SETTINGS_PATH);
+        let database_path = current_dir.join(DATABASE_PATH);
+        let local_player_path = current_dir.join(LOCAL_PLAYERS_PATH);
+        let region_file_path = current_dir.join(REGION_PATH);
+
+        Ok(Self {
+            app_path,
+            current_dir,
+            settings_path,
+            database_path,
+            local_player_path,
+            region_file_path
+        })
+    }
+}
