+++ src-tauri/src/parser/mod.rs
@@ -1,1 +1,2 @@
 pub mod models;
+pub mod data;
