+++ src-tauri/src/parser/models.rs
@@ -5,7 +5,6 @@ use std::str::FromStr;
 
 use bitflags::bitflags;
 use hashbrown::{HashMap, HashSet};
-use lazy_static::lazy_static;
 use log::error;
 use serde::{Deserialize, Deserializer, Serialize};
 use serde_json::{Map, Value};
@@ -724,262 +723,6 @@ pub enum HitFlag {
     DAMAGE_SHARE,
     DODGE_HIT,
     MAX,
-}
-
-lazy_static! {
-    pub static ref NPC_DATA: HashMap<u32, Npc> = {
-        let json_str = include_str!("../../meter-data/Npc.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse NPC data: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref SKILL_DATA: HashMap<u32, SkillData> = {
-        let json_str = include_str!("../../meter-data/Skill.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse SkillData: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref SKILL_EFFECT_DATA: HashMap<u32, SkillEffectData> = {
-        let json_str = include_str!("../../meter-data/SkillEffect.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse SkillEffectData: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref SKILL_BUFF_DATA: HashMap<u32, SkillBuffData> = {
-        let json_str = include_str!("../../meter-data/SkillBuff.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse SkillBuffData: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref COMBAT_EFFECT_DATA: HashMap<i32, CombatEffectData> = {
-        let json_str = include_str!("../../meter-data/CombatEffect.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse CombatEffectData: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref ENGRAVING_DATA: HashMap<u32, EngravingData> = {
-        let json_str = include_str!("../../meter-data/Ability.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse EngravingData: {}", e);
-            HashMap::new()
-        })
-    };
-    pub static ref ESTHER_DATA: Vec<Esther> = {
-        let json_str = include_str!("../../meter-data/Esther.json");
-        serde_json::from_str(json_str).unwrap_or_else(|e| {
-            error!("Failed to parse EstherData: {}", e);
-            Vec::new()
-        })
-    };
-    pub static ref RAID_MAP: HashMap<String, String> = {
-        let json_str = include_str!("../../meter-data/encounters.json");
-        let encounters =
-            serde_json::from_str::<HashMap<String, HashMap<String, Vec<String>>>>(json_str)
-                .unwrap_or_else(|e| {
-                    error!("Failed to parse encounters.json: {}", e);
-                    HashMap::new()
-                });
-        encounters
-            .values()
-            .flat_map(|raid| raid.iter())
-            .flat_map(|(gate, bosses)| bosses.iter().map(move |boss| (boss.clone(), gate.clone())))
-            .collect()
-    };
-    pub static ref GEM_SKILL_MAP: HashMap<u32, Vec<u32>> = {
-        let json_str = include_str!("../../meter-data/GemSkillGroup.json");
-
-        let raw_map: HashMap<String, (String, String, Vec<u32>)> = serde_json::from_str(json_str)
-            .unwrap_or_else(|e| {
-                error!("Failed to parse GemSkillGroup.json: {}", e);
-                HashMap::new()
-            });
-
-        raw_map
-            .into_iter()
-            .filter_map(|(key, entry)| key.parse::<u32>().ok().map(|id| (id, entry.2)))
-            .collect()
-    };
-    pub static ref VALID_ZONES: HashSet<u32> = {
-        let valid_zones = [
-            30801, 30802, 30803, 30804, 30805, 30806, 30807, 30835, 37001, 37002, 37003, 37011,
-            37012, 37021, 37022, 37031, 37032, 37041, 37042, 37051, 37061, 37071, 37072, 37081,
-            37091, 37092, 37093, 37094, 37101, 37102, 37111, 37112, 37121, 37122, 37123, 37124,
-            308010, 308011, 308012, 308014, 308015, 308016, 308017, 308018, 308019, 308020, 308021,
-            308022, 308023, 308024, 308025, 308026, 308027, 308028, 308029, 308030, 308037, 308039,
-            308040, 308041, 308042, 308043, 308044, 308239, 308339, 308410, 308411, 308412, 308414,
-            308415, 308416, 308417, 308418, 308419, 308420, 308421, 308422, 308423, 308424, 308425,
-            308426, 308428, 308429, 308430, 308437, 309020, 30865, 30866,
-        ];
-
-        valid_zones.iter().cloned().collect()
-    };
-    pub static ref SUPPORT_AP_GROUP: HashSet<u32> = {
-        let mut set = HashSet::new();
-        set.insert(101204); // bard
-        set.insert(101105); // paladin
-        set.insert(314004); // artist
-        set.insert(480030); // valkyrie
-
-        set
-    };
-    pub static ref SUPPORT_IDENTITY_GROUP: HashSet<u32> = {
-        let mut set = HashSet::new();
-        set.insert(211400); // bard serenade of courage
-        set.insert(368000); // paladin holy aura
-        set.insert(310501); // artist moonfall
-        set.insert(480018); // valkyrie release light
-
-        set
-    };
-    pub static ref STAT_TYPE_MAP: HashMap<&'static str, u32> = {
-        let mut map = HashMap::new();
-        map.insert("none", 0);
-        map.insert("hp", 1);
-        map.insert("mp", 2);
-        map.insert("str", 3);
-        map.insert("agi", 4);
-        map.insert("int", 5);
-        map.insert("con", 6);
-        map.insert("str_x", 7);
-        map.insert("agi_x", 8);
-        map.insert("int_x", 9);
-        map.insert("con_x", 10);
-        map.insert("criticalhit", 15);
-        map.insert("specialty", 16);
-        map.insert("oppression", 17);
-        map.insert("rapidity", 18);
-        map.insert("endurance", 19);
-        map.insert("mastery", 20);
-        map.insert("criticalhit_x", 21);
-        map.insert("specialty_x", 22);
-        map.insert("oppression_x", 23);
-        map.insert("rapidity_x", 24);
-        map.insert("endurance_x", 25);
-        map.insert("mastery_x", 26);
-        map.insert("max_hp", 27);
-        map.insert("max_mp", 28);
-        map.insert("max_hp_x", 29);
-        map.insert("max_mp_x", 30);
-        map.insert("max_hp_x_x", 31);
-        map.insert("max_mp_x_x", 32);
-        map.insert("normal_hp_recovery", 33);
-        map.insert("combat_hp_recovery", 34);
-        map.insert("normal_hp_recovery_rate", 35);
-        map.insert("combat_hp_recovery_rate", 36);
-        map.insert("normal_mp_recovery", 37);
-        map.insert("combat_mp_recovery", 38);
-        map.insert("normal_mp_recovery_rate", 39);
-        map.insert("combat_mp_recovery_rate", 40);
-        map.insert("self_recovery_rate", 41);
-        map.insert("drain_hp_dam_rate", 42);
-        map.insert("drain_mp_dam_rate", 43);
-        map.insert("dam_reflection_rate", 44);
-        map.insert("char_attack_dam", 47);
-        map.insert("skill_effect_dam_addend", 48);
-        map.insert("attack_power_rate", 49);
-        map.insert("skill_damage_rate", 50);
-        map.insert("attack_power_rate_x", 51);
-        map.insert("skill_damage_rate_x", 52);
-        map.insert("cooldown_reduction", 53);
-        map.insert("paralyzation_point_rate", 54);
-        map.insert("def", 55);
-        map.insert("res", 56);
-        map.insert("def_x", 57);
-        map.insert("res_x", 58);
-        map.insert("def_x_x", 59);
-        map.insert("res_x_x", 60);
-        map.insert("def_pen_rate", 67);
-        map.insert("res_pen_rate", 68);
-        map.insert("physical_inc_rate", 69);
-        map.insert("magical_inc_rate", 70);
-        map.insert("self_shield_rate", 71);
-        map.insert("hit_rate", 72);
-        map.insert("dodge_rate", 73);
-        map.insert("critical_hit_rate", 74);
-        map.insert("critical_res_rate", 75);
-        map.insert("critical_dam_rate", 76);
-        map.insert("attack_speed", 77);
-        map.insert("attack_speed_rate", 78);
-        map.insert("move_speed", 79);
-        map.insert("move_speed_rate", 80);
-        map.insert("prop_move_speed", 81);
-        map.insert("prop_move_speed_rate", 82);
-        map.insert("vehicle_move_speed", 83);
-        map.insert("vehicle_move_speed_rate", 84);
-        map.insert("ship_move_speed", 85);
-        map.insert("ship_move_speed_rate", 86);
-        map.insert("fire_dam_rate", 87);
-        map.insert("ice_dam_rate", 88);
-        map.insert("electricity_dam_rate", 89);
-        map.insert("earth_dam_rate", 91);
-        map.insert("dark_dam_rate", 92);
-        map.insert("holy_dam_rate", 93);
-        map.insert("elements_dam_rate", 94);
-        map.insert("fire_res_rate", 95);
-        map.insert("ice_res_rate", 96);
-        map.insert("electricity_res_rate", 97);
-        map.insert("earth_res_rate", 99);
-        map.insert("dark_res_rate", 100);
-        map.insert("holy_res_rate", 101);
-        map.insert("elements_res_rate", 102);
-        map.insert("self_cc_time_rate", 105);
-        map.insert("enemy_cc_time_rate", 106);
-        map.insert("identity_value1", 107);
-        map.insert("identity_value2", 108);
-        map.insert("identity_value3", 109);
-        map.insert("awakening_dam_rate", 110);
-        map.insert("item_drop_rate", 111);
-        map.insert("gold_rate", 112);
-        map.insert("exp_rate", 113);
-        map.insert("attack_power_addend", 123);
-        map.insert("npc_species_humanoid_dam_rate", 125);
-        map.insert("npc_species_devil_dam_rate", 126);
-        map.insert("npc_species_substance_dam_rate", 127);
-        map.insert("npc_species_undead_dam_rate", 128);
-        map.insert("npc_species_plant_dam_rate", 129);
-        map.insert("npc_species_insect_dam_rate", 130);
-        map.insert("npc_species_spirit_dam_rate", 131);
-        map.insert("npc_species_wild_beast_dam_rate", 132);
-        map.insert("npc_species_mechanic_dam_rate", 133);
-        map.insert("npc_species_ancient_dam_rate", 134);
-        map.insert("npc_species_god_dam_rate", 135);
-        map.insert("npc_species_archfiend_dam_rate", 136);
-        map.insert("vitality", 137);
-        map.insert("ship_booter_speed", 138);
-        map.insert("ship_wreck_speed_rate", 139);
-        map.insert("island_speed_rate", 140);
-        map.insert("attack_power_sub_rate_1", 141);
-        map.insert("attack_power_sub_rate_2", 142);
-        map.insert("physical_inc_sub_rate_1", 143);
-        map.insert("physical_inc_sub_rate_2", 144);
-        map.insert("magical_inc_sub_rate_1", 145);
-        map.insert("magical_inc_sub_rate_2", 146);
-        map.insert("skill_damage_sub_rate_1", 147);
-        map.insert("skill_damage_sub_rate_2", 148);
-        map.insert("resource_recovery_rate", 149);
-        map.insert("weapon_dam", 151);
-        map
-    };
-    pub static ref NPC_GRADE: HashMap<&'static str, i32> = {
-        let mut map = HashMap::new();
-        map.insert("none", 0);
-        map.insert("underling", 1);
-        map.insert("normal", 2);
-        map.insert("elite", 3);
-        map.insert("named", 4);
-        map.insert("seed", 5);
-        map.insert("boss", 6);
-        map.insert("raid", 7);
-        map.insert("lucky", 8);
-        map.insert("epic_raid", 9);
-        map.insert("commander", 10);
-        map
-    };
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
