+++ src-tauri/tauri.conf.json
@@ -1,114 +1,17 @@
 {
   "$schema": "../node_modules/@tauri-apps/cli/schema.json",
+  "productName": "LOA logs",
+  "version": "0.0.1",
+  "identifier": "com.loa-logs",
   "build": {
     "beforeBuildCommand": "npm run build",
     "beforeDevCommand": "npm run dev",
-    "devPath": "http://localhost:5173",
-    "distDir": "../build"
+    "devUrl": "http://localhost:5173",
+    "frontendDist": "../build"
   },
-  "package": {
-    "productName": "LOA Logs",
-    "version": "1.32.5"
-  },
-  "tauri": {
-    "allowlist": {
-      "all": false,
-      "window": {
-        "startDragging": true,
-        "setAlwaysOnTop": true,
-        "setIgnoreCursorEvents": true,
-        "show": true,
-        "hide": true,
-        "maximize": true,
-        "minimize": true,
-        "unminimize": true,
-        "unmaximize": true,
-        "setFocus": true
-      },
-      "clipboard": {
-        "all": true
-      },
-      "shell": {
-        "open": ".+"
-      },
-      "process": {
-        "relaunch": true
-      },
-      "fs": {
-        "readFile": true,
-        "copyFile": true,
-        "scope": ["**"]
-      },
-      "path": {
-        "all": true
-      },
-      "globalShortcut": {
-        "all": true
-      },
-      "http": {
-        "all": true,
-        "request": true,
-        "scope": ["https://*"]
-      }
-    },
-    "bundle": {
-      "active": true,
-      "category": "Utility",
-      "copyright": "GNU General Public License v3.0",
-      "resources": [
-        "meter-data/*.json",
-        "WinDivert.dll",
-        "WinDivert64.sys"
-      ],
-      "icon": [
-        "icons/icon.ico"
-      ],
-      "identifier": "xyz.snow",
-      "longDescription": "LOA Logs - Modern DPS Meter for Lost Ark written in Rust.",
-      "macOS": {
-        "entitlements": null,
-        "exceptionDomain": "",
-        "frameworks": [],
-        "providerShortName": null,
-        "signingIdentity": null
-      },
-      "shortDescription": "LOA Logs",
-      "targets": [
-        "deb",
-        "nsis",
-        "updater"
-      ],
-      "windows": {
-        "webviewInstallMode": {
-          "type": "embedBootstrapper"
-        },
-        "certificateThumbprint": null,
-        "digestAlgorithm": "sha256",
-        "timestampUrl": "",
-        "nsis": {
-          "template": "./nsis/installer.nsi",
-          "license": "../LICENSE",
-          "installerIcon": "icons/icon.ico"
-        }
-      }
-    },
-    "systemTray": {
-      "iconPath": "icons/icon.png",
-      "iconAsTemplate": true
-    },
+  "app": {
     "security": {
       "csp": "default-src *; img-src 'self' https: data:; style-src 'self' 'unsafe-inline'; font-src 'self' data:"
-    },
-    "updater": {
-      "active": true, 
-      "endpoints": [
-        "https://snow.xyz/loa-logs/latest.json"
-      ],
-      "dialog": false,
-      "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IDZEODU5MzNEODY0ODg0REEKUldUYWhFaUdQWk9GYlczc3VSODJPK01RRTlYUER6NzU0eDE1WDYwcmJsY1d4d2EwblZHNlJ4dW8K",
-      "windows": {
-        "installMode": "passive"
-      }
     },
     "windows": [
       {
@@ -138,7 +41,6 @@
         "width": 800,
         "height": 500,
         "decorations": true,
-        "fileDropEnabled": false,
         "visible": false
       },
       {
